ext/ed25519_ref10: Consolidate fe.c and ge.c
This should improve build times
