--- ext/ed25519_ref10/ge_msub.c
@@ -1,11 +1,0 @@
-#include "ge.h"
-
-/*
-r = p - q
-*/
-
-void ge_msub(ge_p1p1 *r,const ge_p3 *p,const ge_precomp *q)
-{
-  fe t0;
-#include "ge_msub.h"
-}
