--- ext/ed25519_ref10/ge_p2_0.c
@@ -1,8 +1,0 @@
-#include "ge.h"
-
-void ge_p2_0(ge_p2 *h)
-{
-  fe_0(h->X);
-  fe_1(h->Y);
-  fe_1(h->Z);
-}
