--- ext/ed25519_ref10/ge_p2_dbl.c
@@ -1,11 +1,0 @@
-#include "ge.h"
-
-/*
-r = 2 * p
-*/
-
-void ge_p2_dbl(ge_p1p1 *r,const ge_p2 *p)
-{
-  fe t0;
-#include "ge_p2_dbl.h"
-}
