--- ext/ed25519_ref10/ge_precomp_0.c
@@ -1,8 +1,0 @@
-#include "ge.h"
-
-void ge_precomp_0(ge_precomp *h)
-{
-  fe_1(h->yplusx);
-  fe_1(h->yminusx);
-  fe_0(h->xy2d);
-}
